config reloader always removes error file before installing configs
So that under no circumstances the error file (or a file with the same
name as the error file) would be copied to the nginx configuration dir.

This could be backward incompatible, but only for cases where there was
no configuration errors.

This is necessary because commit f7f7f16c9dfba8cbc1f390ddfce81de9f8a514ad
excludes the error file from checking for forbidden configurations.