fix mark release detect change
mark-release script had an issue detecting changes. Sometimes there were changes to be picked u for the release, but the script won't notice them, due to how it detected changes.
This PR fixes this by looking for changes in the changelog not only the author section, but with the extra commit lines.

related: kz3 549