don't use internal to check queue status
use notifier.check_events instead of the internal notifier._eventq to
determine if there are other events left in the queue. with this we can
be sure what event is the last event, and then we don't need the sleep
anymore (checking the last event is better than checking the first
event).

cons:
- if we keep touching files with just the right amount of speed we'll
  never reload (but this is rare)

pros:
- no opportunistic sleep
- no using of an internal attribute